refactor: various enhancements.

+++ .files/install/setup-users.sh
@@ -6,14 +6,15 @@ setup_users () {
 
   sudo git clone https://github.com/stoutZero/dotbuntu /dotbuntu
 
-  sudo adduser --gecos "" --shell $(which zsh) pl
-  sudo usermod -aG sudo pl
-  sudo cp -an ~/.{f,g,z}* ~pl
-  [ -d ~.ssh ] && sudo cp -an .ssh ~pl/
-  sudo mkdir ~pl/tmp
-  sudo chown -R pl: ~pl
+  if [ ! -d /home/pl ]; then
+    sudo adduser --gecos "" --shell $(which zsh) pl
+    sudo usermod -aG sudo pl
+    [ -d ~/.ssh ] && sudo cp -an ~/.ssh ~pl/
+    sudo mkdir ~pl/tmp
+    sudo chown -R pl: ~pl
+  fi
 
-  if [ -f /home/pilus/.git/config ]; then
+  if [ -f /home/pl/.git/config ]; then
     sudo su pl -c 'git -C ~/ remote set-url origin /dotbuntu'
   else
     sudo su pl -c 'git config --global --add safe.directory /dotbuntu/.git'
@@ -22,15 +23,21 @@ setup_users () {
     sudo su pl -c 'rm -rf ~/dotbuntu'
   fi
 
-  echo 'pl ALL=(ALL) NOPASSWD:ALL' \
-    | sudo tee /etc/sudoers.d/pl > /dev/null
-  sudo visudo -f /etc/sudoers.d/pl
+  if [ ! -f /etc/sudoers.d/pl ]; then
+    echo 'pl ALL=(ALL) NOPASSWD:ALL' \
+      | sudo tee /etc/sudoers.d/pl > /dev/null
 
-  sudo adduser --gecos "" --shell $(which zsh) deployer
-  sudo cp -an ~/.{f,g,z}* ~deployer
-  [ -d ~.ssh ] && sudo cp -an .ssh ~deployer/
-  sudo mkdir ~deployer/tmp
-  sudo chown -R deployer: ~deployer
+    sudo visudo -f /etc/sudoers.d/pl
+
+    [ $? -eq 1 ] && sudo rm -f /etc/sudoers.d/pl
+  fi
+
+  if [ ! -d /home/deployer ]; then
+    sudo adduser --gecos "" --shell $(which zsh) deployer
+    [ -d ~/.ssh ] && sudo cp -an ~/.ssh ~deployer/
+    sudo mkdir ~deployer/tmp
+    sudo chown -R deployer: ~deployer
+  fi
 
   if [ -f /home/deployer/.git/config ]; then
     sudo su deployer -c 'git -C ~/ remote set-url origin /dotbuntu'
@@ -43,7 +50,7 @@ setup_users () {
 
   sudo chsh -s $(which zsh) root
   sudo cp -an ~/.{f,g,z}* /root/
-  [ -d ~.ssh ] && sudo cp -an .ssh /root/
+  [ -d ~/.ssh ] && sudo cp -an ~/.ssh /root/
   sudo mkdir /root/tmp
   sudo chown -R root: /root/
 
@@ -57,17 +64,17 @@ setup_users () {
   fi
 
   echo 'printing git remote for ~root'
-  sudo cat /root/.git/config
+  sudo su root 'git -C ~ remote get-url origin'
+  echo
+  echo
+
+  echo 'printing git remote for ~deployer'
+  sudo su deployer 'git -C ~ remote get-url origin'
   echo
   echo
 
   echo 'printing git remote for ~pl'
-  sudo cat /home/pl/.git/config
-  echo
-  echo
-
-  echo 'printing git remote for ~pl'
-  sudo cat /home/pl/.git/config
+  sudo su pl 'git -C ~ remote get-url origin'
   echo
   echo
